Add another stack 'mark' type for PostScript procedure braces. These are only used for PostScript calculator functions, but we need them for that. Also, attempting to read a PostScript file as a PDF file meant that these were handled as delimiters, but the PDF interpreter didn't know what to do with them, leading to an infinite loop
Fix the escaping of names; for some reason I had thought that the 2
bytes following the # were decimal, of course they are hex.

We check stream objects to see if their length is correct, but we
should ignore errors unless PDFSTOPNOERROR is set.

Add code to skip any garbage at the start of a file, up to the PDF
header, when repairing files. If we don't find a PDF header when
repairing files, stop and assume the file is too badly broken to repair
(may not even be a PDF file).

Handle the startxref keyword when repairing files

Return an error code in pdf_get_page_dict instead of accidentally
swallowing it.