Change the name of repair_pdf_file to pdf_repair_file to be more consistent with names.
When parsing numbers, if we've already reported a problem don't report
it again.

Permit white space between '<' and a second '<' or hex number.

When we get a closing array bracket, and we have no mark on the stack
just clear the stack and don't return an error (unless PDFSTOPONERROR
is true)

Fix a stupid mistake introduced when we permitted floating point Count
values in page dictionaries, used the wrong dictionary.