If we get an opening string parenthesis '(' while we're already reading a string, don't throw an error (which it should be, this character should be escaped). Instead treat it as nesting and read until the nesting is back to 0.
When repairing a PDF file, if we find an ObjStm stream, then read the
object numbers out of the stream and create xref entries for them.

If when reading a content stream we get an error, ignore any except
ioerror or VMerror and continue reading the content stream.

If we encounter an endobj keyword while reading a content stream
assume that the endstream was missing and close the content stream cleanly