Fix for small glyph distortion
Glyphs with very few pixels (i.e. lower case "i" in 9pt text rendered at 72dpi)
we being distorted due to the specific rules applied by Freetype's dropout
compensation. This isn't a full solution but resolves the change in behaviour
between 2.7.x and 2.9.x (the change was due to different rounding being
applied).

This fix is from upstream:
http://git.savannah.gnu.org/cgit/freetype/freetype2.git/commit/?id=f1458d2e44d8

So should be in future Freetype releases, and not need re-applied here.