Fix a bug in the promotion of an object in the cache to be the MRU, we must not do this if its *already* the MRU.
Fix an off-by-one error when checking the object number to be
dereferenced against the size of the xref table.

If we get a keyword (eg 'obj') when we were expecting an 'endobj' treat
it as a missing endobj and just close the object without complaining
(unless pdfstoponerror is true of course)

When repairing files; if we find an object number > 2^32 - 1 divided
by the size of an xref entry, throw a rangecheck error. Although we hold
object numbers as 64-bit unsigned integers, the memset on 32-bit Windows
builds uses a signed int for the size of memory to set to 0, which can
lead to us failing to clear the memory. If this ever becomes a real
problem we can clear the memory in blocks, but its unlikely that a real
file will ever trip this.

When repairing files, if the offset of the object lies beyond the end
of the file, don't try and read it.

When repairing files, if we reach the end of file trying to read a token
just stop and go on to the next object.

When repairing files, if we find an endstream, don't just assume we
have a preceding dictionary; check there is an object on the stack and
that its a dictionary first.

When repairing files, if the object number we read is less than 1
then simply discard the object, this is not valid.