Prevent use of CID 0 in a PDF/A output file.
Bug #699375 "PDF/A 1b / 2b Validation fails after conversion with Ghostscript on some files"

The input file(s) use CID 0 (which is defined as the /.notdef glyph)
which is illegal in PDF/A files.

'Fixing' this would require us to alter the CID, change the CMap and
the ToUnicode CMap, and runs the obvious risk of the CID we use as a
replacement being used later in the document. We could avoid this by
scanning for 'holes' in the CMap, but even that potentially leaves us
with a problem if there are no empty entries in the CMap.

Since we can't guarantee fixing the problem, a partial fix would be a
great deal of coding, would almost certainly introduce a bug tail and
still wouldn't guarantee a compliant result, don't try. Instead if we
encounter this condition abort PDF/A output, or throw an error if the
PDFACompatibilityPolicy is set that way.