Fix conversion from PDF time to XMP time for pdfwrite
Bug #699182 " Invalid XMP metadata and possibly invalid memory access for dates without timezone in pdfmark"

Two different problems; in the first case a '>' should be '>=' so that
we properly identify an empty string.

Then when converting the string, there was a different off-by-one error
in the buffer index. Also corrected a comment while passing.