Replace EOI marker in JPEG passthrough
Bug #699216 " JPEG passthrough appears to truncate JPEGs in some cases"

The IJG JPEG implementation consumes the EOI marker before the JPEG
passthrough code gets to see it (in general). Most PDF consumers don't
care, but some seem to, so when we terminate the DCT decompression
send the EOI marker to the passthrough code.