First/Last page - Fix image data handling with proc data source
Bug #699215 "Subclassing device leads to infinite(?) loop"

When processing images, its acceptable for the subclassed device to
return 0 (need more data) in response to image_plane_data, if the
source is a file, because the file will reach EOF and the code will
properly terminate.

For procedural (or string) data sources this isn't true. The code needs
to keep track of the number of rows of image data received, and return
1 (completed) when we have received enough data.

Fixed the object filtering device at the same time.