PDF interpreter - fix illegal CIDSystemInfo dictionaries
Bug #699294 " Ghostscript giving errors for attached pdfs and generating blank tiffs"

The supplied example file has several CIDFonts embedded which have
illegal CIDSystemInfo dictionaries; they are missing the required
Registry and Ordering entries.

A long time ago we discarded the CIDSystemInfo from the CIDFont and
created a dummy one, which is why the reporter thinks this 'used to
work'. It didn't. Discarding the information caused problems with the
pdfwrite output.

In this case we simply manufacture the missing data. Its wrong, but it
allows the file to proceed to completion. Naturally we emit a warning
so that users know their file is garbage.