Alter pdf_dict_get() to take a pdf context, so that we can dereference any indirect references and replace the dictionary content with the newly dereferenced object.
Define a new pdf_dict_get_no_store_R which behaves like the old
pdf_dict_get and doesn't store dereferenced objects. We want this for
Resource extraction; to try and avoid any circular references we need
to dereference every object each time we use it. This is used in the
image code.

Update all the uses of pdf_dict_get to pass the context.

Predictor 2 and DCT filters don't work, and hang, so for now don't use
them.

Implement the Do operator, common up the code for Image XObjects and
inline images.

Fix a bug in the object caching; if we promoted a cache entry to be the
MRU entry, but it was already the LRU object, then we would have the
MRU and LRU pointers pointing to the same object, which caused fatal
errors.

When dereferencing compressed objects we weren't setting the object
and generation numbers in the new objects (this happens in parsing
for uncompressed objects). This is vital for caching!

Print warnings when we detect circular references.

When we clear the loop detector back to a mark, if the loop detector
stack is empty, then discard it so we don't have to explicitly discard
the detector when we clear it down.