pdfwrite - drop UTF16-BE DOCINFO strings when creating PDF/A-1
Bug #699392 " 	XMP metadata missing when trying to create PDF/A "

Its not obvious, but it appears to be impossible to embed UTF-16BE
strings in XMP metadata and have them be validated against the same
string in the Info dictionary, because of the specified method for
comparing such strings.

PDF/A-2 states that the document information dictionary should be
ignored, in contrast to the PDF/A-1 specification which says the
XMP and Info dictionary information must be consistent. So this
problem can't arise there.

This commit detects the use of any UTF-16BE string in a DOCINFO
pdfmark and use the usual PDFACompatibilityPolicy to decide whether to
abort PDF/A-1 production, drop the string or raise an error.