Fix #699390 - text position wrong.
The PCL current position (CAP) can be in two states: floating or
fixed.  Floating tracks the origin with margin and orientation changes
fixed does not.  The CAP is floating until we have printable data or a
command that affects the cursor position is received.  We were
mistakenly resetting the the CAP from fixed to float for orientation
changes resulting in incorrect text position.

Norbert Janssen provided the analysis and code change for this fix.