Move installation of standard subclassing devices
Bug #699404 "Failure with psdcmyk -dGraphicsAlphaBits=4 -dFirstPage=1 -dLastPage=1"

The installation of the devices in gdev_prn_open() was slightly too
early, the procs can still (quite reasonably because this is part of
the device installation) be altered by gdev_prn_allocate_memory().

If we had already subclassed the device, then the procs of the
subclassing device (ie firstPage/LastPage) would be altered rather than
the real target device's procs. This led to the wrong graphics library
call, which produced an error.

Moving the subclassing later in the installation solves the problem.