Move the PS Level 1 filters into 'base/'
Oddly, Level 2 filters were already in 'base/', but the Level 1 filters were
still (partially) in 'psi/' (but at least one was partially in base/).

This not only moves the code, but puts both Level 1 and 2 filters into a
'feature' .dev (called psfilters.dev) - on the basis we don't really support
building as a Level 1 only interpreter any more.

It is still up to the interpreter makefile(s) to include the dependency on
psfilter.dev to get these built in.

This is tidier than before, but primarily done so the SubFileDecode filter
is available for use by the in-development C based PDF interpreter.