Bug 699454: 'improve' pdfwrite guesstimate of glyph bbox
When estimating the bounding box of a glyph, we were basically only using the
FontBBox - which is often bogus (as it is in this case).

We'll now, additionally, use the advance width of the glyph in that estimate.

This could possibly result in glyphs being included whose marking operations
are outside the current clip, but whose advance width intersects with the
current clip. But that would be preferable to (as was happening) a glyph inside
the clip being dropped.

(Using the "real" glyph bbox would require executing the complete charstring,
which, obviously, has performance implications)