pdfwrite - when changing ProcessColorModel, don't just change the name
Bug 699490 "Broken PDF/X-3 generated"

The reason that an empty file was generated was due to a mismatch
between the underlying device process color model and the ICC profile
attached to the device. The profile was a CMYK profile, while the
device was set to have 3 components.

This is due to the fact that setting the ProcessColorModel separately
from the ColorConversionStrategy was not actually changing the model,
it just changed the name.

This should, of course, have generated an error and an earlier commit
had already addressed that. In this commit, we call the routine which
actually switches the colour model, and we also count down the ICC
structure attached to the device, and set it to 0. This forces a reload
of the ICC profile. That isn't needed for this particular bug but it
is needed when OutputICCProfile is not set, because we will otherwise
try to use the profile for the prior colour model, which will be
incorrect and leaed to errors.