jbig2dec: Handle text regions without symbol dictionaries.
Normal text regions reference symbol dictionaries. The symbols
in those dictionaries are then referenced to by the text region
and placed somewhere on the page. In the degenerate case a text
region might not have any instances of symbols that it wants to
place on the page. If that is the case then there is no need for
the text region to refer to any symbol dictionaries (since it
doesn't need any symbols). Previously this was treated as a fatal
error, but now this just creates a warning and decoding proceeds.