Increase limit on UniqueIDs for fonts
Bug #699513 "CFF font UniqueID greater then 2^24 - 1 causes error"

The PDF file contains a font with a UniqueID of 0x7fffff which exceeds
the PostScript definition of a font UniqueID (0 -> 2^24 - 1).

Experimentation shows that no PostScript interpreter that I have access
to (other than GS) implements this limit. However, to keep the change
conservative, I'm chosen to raise the limit to 2^32-1 to match the CFF
specification, rather than eliminate it altogether.