Bug 698928: Rework arg_next.
When I worked on arg_next before, I think I had the wrong idea of
what it was doing.

The idea of arg_next (I thought) was that we'd call it, and get back
a pointer to a decoded version of the next argument. Sadly, this
wasn't true - it would leave us with an *encoded* version of the
argument. Now, for most things this didn't matter, for for encodings
like UTF16LE, it certainly did.

So, largely revert the previous misguided commit to this code, and
update arg_next so that it really does return a "decoded" (i.e. always
utf-8 encoded) buffer.

This fixes Bug 698928 where I had an underflow read in a buffer.