Fix a device subclassing problem with colour mapping procedures
Bug #699520 "Seg fault with customer device utilising device subclassing"

When we retrieve colour mapping methods, we need to use the first
subclassing device in the chain which does not have a 'default' for
the colour mapping methods. This is so that the monochrome palette
device in PCL works properly.

In contrast, if none of the subclassing devices overrides the colour
mapping then we want to use the first non-subclassing device's
methods and, importantly, that device as well.