Bug 699379: Fix infinite loop with filenames starting with #.
The argument handling routines treat lines starting with #
as being comments. Unfortunately, the logic for this was failing
to recognise EOF as stopping the search for the end of a line.

Doubly unfortunatately, it was extending this courtesy, intended
for lines within @files to command line arguments too. We take
steps to avoid that too.