Update PageList processing in the FirstPage/LastPage device
The intention of the PageList was that pages would be specified in
ascending order, but the documentation didn't actually say that
explicitly, so update the documentation to make it absolutely clear.

In addition, by specifying a lower page number at the end, it was
possible to cause a buffer overrun, so prevent that here and flag an
error to the user so they know.