pdfwrite - allow PageLabels to be (optionally) transmitted as pdfmarks
Bug #696568 "Treatment of page labels"

Patch supplied by Leon Bottou, the creator and maintainer of the djvu
device for Ghostscript, very slightly modified.

The djvu device wants the PageLabels from existing PDF files sent to
the device as pdfmarks, exactly as they would be from PostScript. The
problem with this from the pdfwrite point of view is that this would
result in us unpicking the number tree to create pdfmark operations, and
then at the pdfwrite end rebuilding a number tree to emit. Not rebuilding
the tree would result in larger output.

This patch checks that the device implements pdfmark, and if it does
checks whether WantsPageLabels is set in the device. If it is then we
assume its pdfwrite (or a similar device prepared to accept the number
tree). If it is not set then we walk the tree and create a new pdfmark
for each page in the tree range.

I'm unable to test this properly, because it relies on a device which
does implement pdfmark, but doesn't want PageLabels. Rudimentary
testing seemed to work as expected, its possible that more complex
examples might require some more work.