device subclassing of patterns, use the correct memory structure
Spotted by Chris while looking at a customer problem, the customer's
code allocs a new pattern accumulator using the device's memory, but it
should really be using the current VM state of the interpreter.

The reason it didn't was because that information wasn't available at
the point where the allocation is made. This commit adds the current
interpreter VM mode to the 'param' structure, which enables us to pass
it to th device, and have the device use it properly.