PDF interpreter improve handling of PageLabels as pdfmark
Patch supplied by Leon Bottou, see Bug #696568. This improves the way
the code handles the number tree when converting it to pdfmark
operations.

I don't have any reasonable way to test this, as it relies on using a
device which implements pdfmark, but doesn't implement WantsPageLabels
(to avoid getting the number tree passed).