Fix subclassing and page buffer reallocation in prn devices
It is possible for us to need to reallocate the page buffer for a
device if (or example) a setpagedevice call results in changed width
or height of the media.

When we do that, if we need to change to (or re-initialise) a clist
then we set the clist 'target' member to be the 'current' device. That
doesn't mean the current device in the graphics state though, if
we have subclassed the device then we will have passed the params
request down the chain, so the rendering device is not the current
device in the gstate.

We need the target of the clist to be the subclassing device, ie the
one in the graphics state, not the 'current' one. Otherwise if we
were to unsubclass the device, then the clist would be pointing at a
device which was freed.

There's no good way to do this from the subclassing code itself, so
I have (reluctantly) modified gdevprn.c to handle this. Any devices
which don't derive from gdevprn, and use the clist, will need to handle
this case themselves.