device subclassing - solve several problems with short-lived subclassing
Nancy's new erasepage optimisation code revealed some interesting
wrinkles in the subclassing code, Firstly;

If we did a save, then inserted a subclassing device, then a restore
we would get a segmentation fault.

This is because when we subclass a device, we make a new copy of the
original device. This device is, of course, created inside the current
save context. When we restore, that restores away the copied device
with the result that the 'child' pointer of the original device is now
pointing at garbage memory.

Fix this by using 'stable' memory instead of immovable memory. Stable
memory is unaffected by save and restore.

This means we also need to 'null' the members of the copied child when
we 'unsubclass' the device. Otherwise garbage collection may end up
trying to enumerate the pointers in the freed device.

Following on from that, when we get a create_compositor call in the
subclassing code we do a dance whereby we patch the method in the
compositor handler that gets pushed, so that it points to the subclass
device, not the device's ultimate child. Of course, when we unsubclass
the device, we need to unpatch the compositor as well.

That one may need more work.

Finally (for now) we weren't properly handling the parent and child
pointers for the list when we removed a device from the middle of a
chain. The doubly-linked pointers weren't being properly updated.