pdfwrite - fix stored device pointers in streams, when subclassing
Yet more evidence of nasty hackery in pdfwrite. It stores a copy of the
device pointer in the pdf stream_state. Of course, if we had a
subclassed device when we created the stream, and unsubclassed it before
we closed the (eg if we create a high level pattern before the first
marking operation) then the stored pointer was pointing at the wrong
(child, freed) device.

Just do some pointer traversal to make sure we store the top level one
and use the bottom level one when we need it.