Fix the display device with subclassing devices
When the display device is chosen, the executable calls display_set_callbacks
on the current device. That code then sets a member in the current
device, *without* checking the device is actually the display device.

It does this *before* opening the device....

This is a quick hack to resolve the problem by descending to the bottom
child device if there is a chain of devices in place. We should really
check the device name at each step.

In addition the Windows display device keeps a pointer to the device along
with each 'image'. Not sure why, but it breaks if we are using device
subclassing, because the device can change.

Finally; the code doesn't cope with being unable to find a matching
image with the same device. It sets the 'img' to NULL and then goes ahead
and tries to use it anyway. Fix that at the same time so we don't crash!