ps2write - improve handling of Separation colour space with Lab alternate
Bug #699535 "Separation Colors with alternate Color Lab are not converted correctly when device (e)ps2write is used"

We can't embed a Lab space in the output from ps2write with the current
colour handling. So instead spot the condition and convert the
/Separation alternate space into the underlying ProcessColorModel.

This preserves the /Separation ink, though the resulting tint transform
is less than ideal.

At the same time, make sure we don't try to handle the ICC Lab space
as if it were a device space. Instead return a rangecheck error so that
we convert the whole space into the ProcessColorModel.