jbig2dec: Print warning messages even for non-debug builds.
This is useful because jbig2dec emits warnings when it encounters
broken files it is able to cope with, e.g. referrals to symbols
outside the symbol dictionary are ignored but a warning is
emitted. Ghostscript ought to output warnings in these cases so
that users know that the final render might not be the intended
one because the input file is broken.