PDF interpreter - improve handling of 'q' inside a text object
Bug #699551 "Text with rendering mode stroke is not displayed"

We already did deal with a q/Q pair inside a text object (BT/ET pair)
but a couple of points were missed. We didn't check for the existence
of a qTextSavematrix in settextstate, which meant that we didn't
set the text matrix. So vector and image marking worked, but text did
not.

We also didn't check for the existence of a qSaveTextMatrix when handling
the stroke width for a text rendering mode which involves stroking the
text.

This commit exhibits one 'progression' (ie it now matches Acrobat) in
the Quality Logice PDF 2.0 FTS files. I intend to raise this with QL as
the file is demonstrably invalid.

It also adversely impacts the performance of text in PDF files, since we
must now do additional checking on every text operation. This has pushed
a small number of our test files beyond the point where they start to
time out on our automated testing.