Fix tiffpack and tifflzw devices
Commit 0ef309305417776a9f8b8cd143fbf3c5e89456b2 moved the FillOrder
parameter into the 'generic' tiff fax device structure, and added a new
BlackIs1 parameter.

The macro FAX_DEVICE_BODY was updated to reflect this but unfortunately
two of the devices, tiffpack and tifflzw do not use the FAX_DEVICE_BODY
macro, as they need to use device-specific code for the output_page
device method.

This meant that the initialisation of these devices was incorrect and
led to failures in our automated testing.