pdfwrite - don't retry CharProc capture on error.
Bug #699561 "Infinite loop or Segfault"

The input PostScript file for this case is invalid and throws a
rangecheck error with normal devices.

However when the output device is pdfwrite in the case of *any* error
handling text we fall back to the 'default implementation' which means
we run the glyph description normally, and capture the bitmap into a
type 3 bitmap font.

In the case of this file, however, when we try to use the captured
bitmap we again throw an error. This causes us to run the glyph
description again and capture the bitmap. Of course, that again fails
when we try to use it leading to an infinite loop. But not any simple
loop, it involves exiting the code and re-entering it after capture.

This commit adds a new variable, captured_pte_index. When we complete
a capture, we record the index (the pointer to the input character codes)
from the text enumerator. We then try to process the text as normal. If
we get an error, before falling back to the default implementation, we
test the current text enumerator index. If its the same as the captured
index then we know that it was the captured CharProc threw an error.

In the case of this kind of error we simply return the error, there is
nothing further we can do.