Device Subclassing - fix memory leaks with child device
We should not patch out the finalize routine (now that we have a custom
one), that prevents the child device finalize being run, leading to
memory leaks with ICC profiles.

Also, now we have a finalize routine, we need to free the child device
if there is one present. The child device's own finalize routine will
clean everything up.