Device subclassing - synchronise devices on put_params
Commit 434fb65601b91e14fe4f846dcbb92a68d939faf9 on the erasepage
branch fixed the initial problem, but was not a sufficient solution.

While it seems correct to update the 'target' of the clist so that it
points to the head of the subclassing chain, we can run into problems
using that device, if its parameters (eg width and height) do not
match the ultimate device.

This can happen if we get a put_params() which alters the device
parameters. Although we update the chain after the put_params has
completed, gdevprn-derived devices will reallocate memory or
re-initialise the clist *during* the course of the put_params. If we
don't resynchronise the chain but do update the clist target, then the
clist re-initialisation will be using the wrong parameters.