Device Subclassing - fix memory leak of the 'stype' device member
This is all horribly crufty and really not possible to explain.

With the addition of the custom finalize() for subclassing devices we
sorted this out for the case where all the devices were garbage
collected (the GC, the custom finalize and the device finalize take
care of all the wrinkles here).

But for unsubclassing we cannot use the finalize() routine for the
original device when freeing the child. For most devices this does
actually work, but if the device itself has a custom finalize method
(rare, but see psdcmyk) then if we allow the finalize routine to run
it might (and in the case of psdcmyk does) free memory which is still
being pointed at by the parent (original) device structure.

So we need to NULL the finalize routine in the stype structure, tell
the memory manager to use the newly modified structure, then free the
child device memory and finally free the stype structure.

We also need to properly handle the reference counts of the icc_struct
and PagesList objects, but we had done that already in previous commits.

For safety, set the parent and child pointers to NULL and set the child
device structure reference count to 0 to ensure it is not retained.

Finally; there was a minor error when patching back the memory manager
'stype' copy in the parent device, fix that now too.