pdfwrite - improve further the glyph bounding box calculation
Bug #699571 "Missing characters when convert eps to pdf. Ok when convert to png."

When deciding whether to include a text string in the output PDF file
we compare the bounding box of the string with the bounding box of the
current clip path. If the string bbox is not at least partially within
the clip, then we don't emit it.

To do this we previously used the font's FontBBox in order to estimate
the size of each glyph in the string.

Commit 6a4202691c4317a698fa23c5309ad8974fe2b003 fixed a bug where the
font has a wildly incorrect FontBBox leading to us deciding that the
glyph was outside the clip area.

In that commit Chris Liddell noted that using the 'real' glyph bounding
box would require executing the complete CharString which has
performance implications.

Unfortunately this latest bug has forced us into this course. The EPS
file draws the same glyph multiple times, clipping portions of it each
time. Sometimes it uses charpath/fill to render the clipped portion, and
sometimes it uses an xyshow where the advance in both directions is 0.
At the same time, the font has a completely inaccurate FontBBox.

Because the advance width is 0 the previous fix does not address this
situation. The FontBBox is so inaccurate that most of the times the
glyph is rendered it is deemed to be outside the clip.

The only remaining solution is to determine the *actual* Bounding Box
of the glyph, which means we have to execute the glyph description.

This does result in some degradation in performance.