PDF interpreter - Remove old annotation handling for pdfwrite
Bug #699619 "Annotations included twice with pdfwrite"

At some point in the past we handled a very few annotations with
pdfwrite and did so in the pdfshowpage_finish routine. Since then we
have moved the handling of annotations for pdfwrite into the
showpagecontents routine, along with regular annotation rendering.

But leaving the old code in place meant that some annotations (/Link
and /Text) were being processed twice (for pdfwrite). Additionally
the old code didn't honour the ShowAnnots or PreserveAnnots controls.

Just remove the old code.