PDF interpreter - resurrect Link handling improvements
commit 61f53ab8f9e586c8722ccd0ea7de3722b0d40cd2 removed the duplication
of /Link and /Text annotation handling, and used the newer code for
handling these instead.

It transpires that the /Link handling in the old (removed) code was
considerably more functional than that in the newer code...

Here we replace the new code with the old code in order to get the same
functionality, but with the proper controls in place, and only handling
annotations in one place instead of 2.