Bug 690131: Add epo subclass device to do erasepage optimization.
This optimization only works for devices that have a fillpage
implementation that is "gx_default_fillpage".  For other devices the
subclass device will not be installed.  If the device transitions to
clist (or any other non-gx_default_fillpage device), we will stop
doing the optimization.

The optimization basically swallows(defers) all fillpages, remembering
the color (which is probably always white, but it will work with any
"pure" color).  Then when the first actual marking operation happens,
we do a fill_rectangle operation with the remembered color.