Refactor colour space handling
Major reworking of the way that colour spaces are handled (other than
the  color+space operators, like 'rg').

We now set colour spaces from names (eg /DeviceRGB or /CS1) or from
an array. If a name is not a standard name, then we look it up in the
Resources dictionary and proceed from there. This makes the whole
colour space handling recursive.

Fixed a number of problems with the existing code in terms of the use
of gs_swapcolors(), <= checks that should have been < and so oon.

Added code to handle ICCBased colour spaces.

Currently the code will handle Gray, RGB, CMYK, ICCBased, Lab, CalGray
and CalRGB spaces. Code for Indexed spaces is incomplete, still need
/Separation, /DeviceN and /Pattern.